apps/backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
SQLAlchemy==2.0.49
aiosqlite==0.22.1
greenlet==3.4.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
slowapi==0.1.9
python-dotenv==1.2.2
openai==2.31.0
requests==2.33.1
PyYAML==6.0.3
